Top merge, ignore clk N, start test
The top module has been merged back together into one file. The blocks have been made combinational to allow both edges of the clk to be detected, allowing for a 50Mhz clk/toggle. When making this change, there was an error for using the reset as a clock, the constraint was changed to ignore this condition. The simulaiton was able to start, but not able to provide an input to the keypad signal.